# Conflicts:
#	client/src/app/clients/client-list/client-list.component.ts
#	client/src/app/users/user-list/user-list.component.ts

+++ server/src/test/java/com/itechart/trucking/dao/impl/JpaItemDaoTest.java
@@ -36,7 +36,7 @@ public class JpaItemDaoTest {
 
     private static final String NAME="item_name";
 
-    private static final Integer PRICE= 15;
+    private static final Float PRICE= 15f;
 
     private static final ItemUnit UNIT = new ItemUnit("PCS");
